// autogpt_platform/frontend/src/app/api/__generated__/models/agentInfo.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { AgentInfoCreator } from "./agentInfoCreator";
import type { AgentInfoCategory } from "./agentInfoCategory";
import type { AgentInfoRating } from "./agentInfoRating";
import type { AgentInfoRuns } from "./agentInfoRuns";
import type { AgentInfoIsFeatured } from "./agentInfoIsFeatured";
import type { AgentInfoStatus } from "./agentInfoStatus";
import type { AgentInfoCanAccessGraph } from "./agentInfoCanAccessGraph";
import type { AgentInfoHasExternalTrigger } from "./agentInfoHasExternalTrigger";
import type { AgentInfoNewOutput } from "./agentInfoNewOutput";
import type { AgentInfoGraphId } from "./agentInfoGraphId";
import type { AgentInfoGraphVersion } from "./agentInfoGraphVersion";
import type { AgentInfoInputSchema } from "./agentInfoInputSchema";
import type { AgentInfoOutputSchema } from "./agentInfoOutputSchema";
import type { AgentInfoInputs } from "./agentInfoInputs";

/**
 * Information about an agent.
 */
export interface AgentInfo {
  id: string;
  name: string;
  description: string;
  /** marketplace or library */
  source: string;
  in_library?: boolean;
  creator?: AgentInfoCreator;
  category?: AgentInfoCategory;
  rating?: AgentInfoRating;
  runs?: AgentInfoRuns;
  is_featured?: AgentInfoIsFeatured;
  status?: AgentInfoStatus;
  can_access_graph?: AgentInfoCanAccessGraph;
  has_external_trigger?: AgentInfoHasExternalTrigger;
  new_output?: AgentInfoNewOutput;
  graph_id?: AgentInfoGraphId;
  graph_version?: AgentInfoGraphVersion;
  /** JSON Schema for the agent's inputs (for AgentExecutorBlock) */
  input_schema?: AgentInfoInputSchema;
  /** JSON Schema for the agent's outputs (for AgentExecutorBlock) */
  output_schema?: AgentInfoOutputSchema;
  /** Input schema for the agent, including field names, types, and defaults */
  inputs?: AgentInfoInputs;
}