// autogpt_platform/frontend/src/app/api/__generated__/models/agentsFoundResponse.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { ResponseType } from "./responseType";
import type { AgentsFoundResponseSessionId } from "./agentsFoundResponseSessionId";
import type { AgentInfo } from "./agentInfo";

/**
 * Response for find_agent tool.
 */
export interface AgentsFoundResponse {
  type?: ResponseType;
  message: string;
  session_id?: AgentsFoundResponseSessionId;
  title?: string;
  agents: AgentInfo[];
  count: number;
  name?: string;
}